Address PR #780 feedback: extract FolderToolbar and add design tokens
- Extract FolderToolbar component from FolderDocumentBrowser (1165→704 lines)
- Add danger/success semantic color tokens to osLegalStyles.ts
- Replace hardcoded #dc2626 colors with OS_LEGAL_COLORS.danger tokens
- Add mobile folder title indicator for better navigation context
- Add data-testid attributes to view toggle buttons for testing
- Skip MetadataWorkflow test pending extensive mock infrastructure updates
